separation of MascotsListener + revive mascot + revive cost ...
